Move ErrDecodingCredential from configdir to credentials package
