+++ pkg/secrethub/configdir/dir.go
@@ -19,15 +19,6 @@ var (
 	ErrCredentialNotFound = errors.New("credential not found")
 )
 
-type ErrDecodingCredential struct {
-	Location string
-	Err      error
-}
-
-func (e ErrDecodingCredential) Error() string {
-	return fmt.Sprintf("error decoding credential loaded from '%s': %v", e.Location, e.Err)
-}
-
 // Dir represents the configuration directory located at some path
 // on the file system.
 type Dir struct {
@@ -119,7 +110,7 @@ func (f *CredentialFile) Read(decoder credentials.KeyDecoder) (credentials.Key, 
 	}
 	key, err := decoder.Decode(bytes)
 	if err != nil {
-		return credentials.Key{}, ErrDecodingCredential{
+		return credentials.Key{}, credentials.ErrDecodingCredential{
 			Location: f.path,
 			Err:      err,
 		}
